requirements.txt
pygame-ce==2.5.1
pygame_gui==0.6.12
python-i18n==0.3.9
